feat: add decimal support

+++ src/main/java/com/gbazuev/TokenStorage.java
@@ -1,16 +1,17 @@
 package com.gbazuev;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 public class TokenStorage {
     private ArrayList<Character> operators = new ArrayList<>();
-    private ArrayList<Integer> operands = new ArrayList<>();
+    private ArrayList<BigDecimal> operands = new ArrayList<>();
 
     public void setOperatorsList(ArrayList<Character> operators) {
         this.operators = operators;
     }
 
-    public void setOperandsList(ArrayList<Integer> operands) {
+    public void setOperandsList(ArrayList<BigDecimal> operands) {
         this.operands = operands;
     }
 
@@ -18,20 +19,20 @@ public class TokenStorage {
         return operators;
     }
 
-    public ArrayList<Integer> getOperandsList() {
+    public ArrayList<BigDecimal> getOperandsList() {
         return operands;
     }
 
-    public int[] getOperands() {
+    public BigDecimal[] getOperands() {
         int currentOperatorIndex = getOperatorIndex();
-        return new int[]{operands.get(currentOperatorIndex), operands.get(currentOperatorIndex + 1)};
+        return new BigDecimal[]{operands.get(currentOperatorIndex), operands.get(currentOperatorIndex + 1)};
     }
 
     public char getOperator() {
         return operators.get(getOperatorIndex());
     }
 
-    public void replaceOperands(int result) {
+    public void replaceOperands(BigDecimal result) {
         int currentOperatorIndex = getOperatorIndex();
         operands.remove(currentOperatorIndex);
         operands.remove(currentOperatorIndex);
